+++ gen/generator.go
@@ -122,6 +122,7 @@ type YamlConfig struct {
 	BulkSizeCreate           int                   `yaml:"bulk_size_create"`
 	ImportNameQuery          bool                  `yaml:"import_name_query"`
 	AdjustBody               bool                  `yaml:"adjust_body"`
+	DeprecationMessage       string                `yaml:"deprecation_message"`
 }
 
 type YamlConfigAttribute struct {